apps/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg[binary]>=3.2.4,<3.3
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.1
feedparser==6.0.11
python-dateutil==2.9.0.post0
numpy>=2.0.0,<2.2.0
scikit-learn==1.5.2
